+++ libsolidity/formal/Z3CHCInterface.cpp
@@ -30,9 +30,13 @@ Z3CHCInterface::Z3CHCInterface():
 	m_solver(*m_context)
 {
 	// This needs to be set globally.
-	z3::set_param("rewriter.pull_cheap_ite", true);
+//	z3::set_param("rewriter.pull_cheap_ite", true);
 	// This needs to be set in the context.
 	m_context->set("timeout", queryTimeout);
+
+	z3::params p(*m_context);
+	p.set("fp.spacer.q3.use_qgen", true);
+	m_solver.set(p);
 }
 
 void Z3CHCInterface::declareVariable(string const& _name, Sort const& _sort)
@@ -82,8 +86,10 @@ pair<CheckResult, vector<string>> Z3CHCInterface::query(Expression const& _expr)
 			break;
 		}
 		case z3::check_result::unknown:
+		{
 			result = CheckResult::UNKNOWN;
 			break;
+		}
 		}
 		// TODO retrieve model / invariants
 	}